Airline Analysis
In this project, you'll imagine that you work for a travel agency and need to know the ins and outs of airline prices for your clients. You want to make sure that you can find the best deal for your client and help them to understand how airline prices change based on different factors.

You decide to look into your favorite airline. The data include:

miles: miles traveled through the flight
passengers: number of passengers on the flight
delay: take-off delay in minutes
inflight_meal: is there a meal included in the flight?
inflight_entertainment: are there free entertainment systems for each seat?
inflight_wifi: is there complimentary wifi on the flight?
day_of_week: day of the week of the flight
weekend: did this flight take place on a weekend?
coach_price: the average price paid for a coach ticket
firstclass_price: the average price paid for first-class seats
hours: how many hours the flight took
redeye: was this flight a redeye (overnight)?
In this project, you'll explore a dataset for the first time and get to know each of these features. Keep in mind that there's no one right way to address each of these questions. The goal is simply to explore and get to know the data using whatever methods come to mind.

You will be working in this file. Note that there is the file Airline Analysis_Solution.ipynb that contains the solution code for this project. We highly recommend that you complete the project on your own without checking the solution, but feel free to take a look if you get stuck or if you want to compare answers when you're done.